initial scheme preparation for AWS EC2 integration
