+++ pom.xml
@@ -62,6 +62,11 @@
 			<artifactId>spring-cloud-starter-vault-config</artifactId>
 			<version>2.2.0.RELEASE</version>
 		</dependency>
+		<dependency>
+			<groupId>org.springframework.cloud</groupId>
+			<artifactId>spring-cloud-aws-context</artifactId>
+			<version>2.2.0.RELEASE</version>
+		</dependency>
 		<!-- Jackson due to SpringBootStarterParent dependency problems -->
 		<dependency>
 			<groupId>com.fasterxml.jackson.core</groupId>
