+++ src/main/java/io/mixeway/config/Constants.java
@@ -269,4 +269,6 @@ public class Constants {
     public static final String VULN_TYPE_NETWORK = "Network";
     public static final String VULN_TYPE_OSPACKAGE = "OSPackage";
     public static final String STATUS_REMOVED = "REMOVED";
+    public static final String IAAS_API_TYPE_AWS_EC2 = "AWS EC2";
+	public static final String IAAS_API_TYPE_OPENSTACK = "OpenStack";
 }
