+++ src/main/java/io/mixeway/db/repository/IaasApiRepository.java
@@ -5,7 +5,11 @@ import org.springframework.stereotype.Repository;
 
 import io.mixeway.db.entity.IaasApi;
 
+import java.util.List;
+
 @Repository
 public interface IaasApiRepository extends JpaRepository<IaasApi,Long>{
 
+    List<IaasApi> findByStatus(Boolean status);
+    List<IaasApi> findByStatusAndEnabled(Boolean status, Boolean enabled);
 }
