+++ src/main/resources/db/changelog/db.changelog-master.sql
@@ -1068,3 +1068,12 @@ update codeproject set skipallscan=true where skipallscan is null;
 --changeset siewer:197
 alter table nessusscan add column retries int;
 update nessusscan set retries=0;
+
+--changeset siewer:198
+create table iaasapitype (
+    id serial primary key,
+    name text
+);
+insert into iaasapitype (name) values ('OpenStack'), ('AWS EC2');
+alter table iaasapi add column iaasapitype_id int references iaasapitype(id) on delete cascade ;
+update iaasapi set iaasapitype_id = 1;
